rest arrayrun endpoint switched from requisition controller to array run controller
