+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/rest/ArrayRunRestController.java
@@ -3,8 +3,10 @@ package uk.ac.bbsrc.tgac.miso.webapp.controller.rest;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,8 +22,10 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import jakarta.servlet.http.HttpServletRequest;
 import uk.ac.bbsrc.tgac.miso.core.data.Array;
 import uk.ac.bbsrc.tgac.miso.core.data.ArrayRun;
+import uk.ac.bbsrc.tgac.miso.core.data.Sample;
 import uk.ac.bbsrc.tgac.miso.core.service.ArrayRunService;
 import uk.ac.bbsrc.tgac.miso.core.service.ArrayService;
+import uk.ac.bbsrc.tgac.miso.core.service.SampleService;
 import uk.ac.bbsrc.tgac.miso.core.util.LimsUtils;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginatedDataSource;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginationFilter;
@@ -44,6 +48,12 @@ public class ArrayRunRestController extends AbstractRestController {
 
   @Autowired
   private AdvancedSearchParser advancedSearchParser;
+
+  @Autowired
+  private SampleService sampleService;
+
+  @Value("${miso.detailed.sample.enabled}")
+  private Boolean detailedSample;
 
   private final JQueryDataTableBackend<ArrayRun, ArrayRunDto> jQueryBackend =
       new JQueryDataTableBackend<ArrayRun, ArrayRunDto>() {
@@ -72,6 +82,56 @@ public class ArrayRunRestController extends AbstractRestController {
       throws IOException {
     return jQueryBackend.get(request, advancedSearchParser, PaginationFilter.project(id));
   }
+
+  // idk if the full URL is needed or something, and I think the produces part is fine but not sure
+  @GetMapping(value = "/dt/requisition/{requisitionId}", produces = "application/json")
+  public @ResponseBody DataTablesResponseDto<ArrayRunDto> dataTableByRequisition(@PathVariable long requisitionId,
+      HttpServletRequest request)
+      throws IOException {
+
+
+    // I don't think I need the request itself to be honest
+    // return jQueryBackend.get(request, advancedSearchParser,
+    // PaginationFilter.requisitionId(requisitionId));
+
+    List<Long> allSamples = new ArrayList<Long>();
+
+    // requisitioned
+    List<Long> sampleIds = sampleService.list(0, 0, false, null,
+        PaginationFilter.requisitionId(requisitionId)).stream().map(Sample::getId)
+        .collect(Collectors.toCollection(() -> new ArrayList<>()));
+
+    // supplemental
+    sampleIds.addAll(sampleService.list(0, 0, false, null,
+        PaginationFilter.supplementalToRequisitionId(requisitionId))
+        .stream().map(Sample::getId).toList());
+
+    if (detailedSample) {
+      allSamples = arrayRunService.getSamplesDescendantslList(sampleIds, requisitionId);
+      // if in detailed mode, get the aliquot samples and any aliquot descendants
+    } else {
+      allSamples = sampleIds; // if not detailed mode, then just get the sample IDs
+    }
+
+    List<ArrayRun> arrayRuns = arrayRunService.listBySamplesIds(allSamples);
+
+    DataTablesResponseDto<ArrayRunDto> dtResponse = new DataTablesResponseDto<>(); // this is done explicitly here, as
+                                                                                   // we cannot use paginationFilter to
+                                                                                   // filter array run by requisition ID
+                                                                                   // directly
+
+    dtResponse.setITotalRecords((long) arrayRuns.size());
+    dtResponse.setITotalDisplayRecords((long) arrayRuns.size());
+    dtResponse.setAaData(arrayRuns.stream().map(Dtos::asDto).collect(Collectors.toList()));
+    dtResponse.setSEcho(Long.valueOf(request.getParameter("sEcho")));
+    // ignore setting the error message for now
+    return dtResponse;
+
+
+
+  }
+
+
 
   @PostMapping(produces = "application/json")
   @ResponseStatus(HttpStatus.CREATED)
